requirements.txt
dnspython ~= 2.3.0
pytest >= 6.0.0
pytest-cov ~= 4.0.0
mypy ~= 1.2.0
requests
requests_cache
flask
pyyaml